New images after bug fix
The code was running way too slowly, and the professor saw that the call to activate the worksheet was within the loop, and thus getting called every pass through the loop (>3000 times!!). Once the call was moved outside the for loop, the run time for the code was reduced from 1.8 seconds to .11 seconds! So, anyway, after this fix, new images were needed for the analysis and here they are.